requirements.txt
awscli==1.18.86
boto3==1.14.9
ipython==7.15.0
jupyter==1.0.0
requests==2.24.0
PyYAML==5.3.1
